Refactor : enum 패키지 rename
